pyproject.toml
[tool.poetry]
name = "otter-point"
version = "0.1.0"
description = ""
authors = ["Ken Hung <cchung1985@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.3"
uvicorn = "^0.29.0"
web3 = "^6.18.0"
psycopg = {extras = ["pool"], version = "^3.1.18"}
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
httpx = "^0.27.0"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
